backend/requirements.txt
Flask>=3.0.0
Flask-SQLAlchemy>=3.1.0
Flask-Migrate>=4.0.0
Flask-JWT-Extended>=4.6.0
Flask-CORS>=4.0.0
Flask-Limiter>=3.5.0
python-dotenv>=1.0.0
psycopg2-binary>=2.9.0
PyMuPDF>=1.23.0
python-docx>=1.1.0
google-generativeai>=0.3.0
sentence-transformers>=2.2.0
numpy>=1.24.0
scikit-learn>=1.3.0
werkzeug>=3.0.0
celery>=5.3.0
redis>=5.0.0
setuptools>=65.0.0
wheel>=0.40.0
